refactor: move registerPlayerContext
